requirements.txt
tensorflow==2.15.0
flask==3.0.3
pillow==10.3.0